Surface: SurfaceListener::onBufferDetached now returns buffer id
Previously returned slot, which wasn't partuclarly useful. As
evidenced by how it wasn't used anywhere.

Bug: 486182281
Flag: EXEMPT PURE_REFACTOR
Test: builds

+++ services/surfaceflinger/DisplayHardware/VirtualDisplay/SinkSurfaceHelper.h
@@ -109,7 +109,7 @@ public:
     virtual void onBufferReleased() override;
     virtual void onRemoteDied() override;
     virtual void onBuffersDiscarded(const std::vector<sp<GraphicBuffer>>&) override {}
-    virtual void onBufferDetached(int) override {}
+    virtual void onBufferDetached(uint64_t) override {}
 
 private:
     void connectSinkSurfaceTask(std::shared_ptr<std::promise<SinkSurfaceData>> promise);
